flaskProject/requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
easyocr==1.4.2
feedparser==6.0.8
Flask==2.0.3
fonttools==4.33.3
idna==3.3
imageio==2.19.0
imutils==0.5.4
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tifffile==2022.4.28
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
